feat(RacingGame): `controller` 생성
