+++ src/main/java/race/RacingCarApplication.java
@@ -3,6 +3,7 @@ package race;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import race.contorller.RacingGame;
 import race.domain.Car;
 import race.domain.Cars;
 import race.view.InputView;
@@ -14,24 +15,6 @@ import utils.number.RandomNumberGenerator;
 public class RacingCarApplication {
 
     public static void main(String[] args) {
-        String[] carNames = InputView.requestCarNames();
-        int countOfTrial = InputView.requestCountOfTrial();
-
-        ResultView.start();
-        Cars cars = Cars.createInstance(carNames);
-
-        NumberGenerator randomGenerator = new RandomNumberGenerator();
-        racing(cars, randomGenerator, countOfTrial);
-
-
-        ResultView.showWinners(cars.chooseWinners());
-    }
-
-    private static void racing(Cars cars, NumberGenerator numberGenerator, int countOfTrial) {
-        while (countOfTrial > 0) {
-            cars.moveCars(numberGenerator);
-            ResultView.showProcess(cars);
-            countOfTrial--;
-        }
+        RacingGame.run();
     }
 }
